added flag for overall pass / fail in solution-test
