+++ cmd/solution/types.go
@@ -338,51 +338,3 @@ func getFmmEventsFromFile(filePath string) []*FmmEvent {
 	}
 	return fmmEvents
 }
-
-type SolutionTestObjects struct {
-	Tests         []SolutionTestObject `json:"tests"`
-	InitialDelay  int                  `json:"initialDelay,omitempty"`
-	MaxRetryCount int                  `json:"retryCount,omitempty"`
-	RetryDelay    int                  `json:"retryDelay,omitempty"`
-}
-
-type SolutionTestObject struct {
-	Name        string                  `json:"name,omitempty"`
-	Type        string                  `json:"type,omitempty"`
-	Description string                  `json:"description,omitempty"`
-	Setup       SolutionTestSetup       `json:"setup"`
-	Assertions  []SolutionTestAssertion `json:"assertions"`
-}
-
-type SolutionTestSetup struct {
-	Type     string      `json:"type"`
-	Input    interface{} `json:"input,omitempty"`
-	Location string      `json:"location,omitempty"`
-}
-
-type SolutionTestAssertion struct {
-	UQL        string                           `json:"uql"`
-	Transforms []SolutionTestAssertionTransform `json:"transforms"`
-}
-
-type SolutionTestAssertionTransform struct {
-	Type       string `json:"type"`
-	Expression string `json:"expression,omitempty"`
-	Message    string `json:"message,omitempty"`
-	Location   string `json:"location,omitempty"`
-}
-
-type SolutionTestResult struct {
-	ID string `json:"testRunId"`
-}
-
-type SolutionTestStatusResult struct {
-	Complete       bool            `json:"completed"`
-	StatusMessages []StatusMessage `json:"statusMessages"`
-}
-
-type StatusMessage struct {
-	Timestamp string   `json:"timestamp"`
-	Message   string   `json:"message,omitempty"`
-	Statuses  []string `json:"statuses,omitempty"`
-}
